+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Improvement:
 - ValidationError.detail now accept optional 2nd parameter `allowUnsafeValidatorDetails`
 - macro: add `introspect`
 - prevent redundant route compilation
+- merge multiple macro resolve response
 
 Bug fix:
 - [#1537](https://github.com/elysiajs/elysia/issues/1537) websocket: ping/pong not being called